Working Build
Appears to have all Features working including a GUI that flashes a purple outline then fades out to alert you that it moved a shapekeys weight down to 0 and Set the basis as the active key. Also includes and N menu Ability to turn the flash off aswell as overall shutoff of the tool.